MHV-53916 update test location
